* Hometimeline working * Cleaned up Main Timeline Activity
